+++ changelog.md
@@ -2,6 +2,7 @@
 * Added back Farming features: Crops, Bone Meal, Livestock, Hoes
   * Rich farmland has been disabled by default
   * Crops can now grow with bone meal when farmland is not moist
+* Added back Item features: Unvanishable Items (aka Unbreakable Items)
 * Fixed Water Fall Damage feature not working properly
 
 ## 5.2.0.0-alpha